Added commands
Commands are for users to do special things outside of just chatting.
Right now that's just emotes and nicknames. The commands now need to be
initialized at the beginning of execution. These commands are handled as
function pointers in a command structure that then gets attached to a
ptab for later retrieval.