Command parsing and a simpler sending functions
First we command parsing. No use in having commands if we can't
determine if we're looking at a command at all. This meant that we
needed to pull some functionality out of our send functions and put them
into the parsing. For instance, the output should be determined by the
function creating the buffer to send, rather than the send function
itself.